+++ docs/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 * Updated dependency versions.
 * Improved the error message when Docker or Vagrant isn't installed.
-* Fixed a bug that caused the spinner to continue running if the environment is a docker container that cannot be found.
-* Improved error handling in the case a docker container cannot be found.
+* Fixed a bug that caused the spinner to continue running if the environment is a Docker container that cannot be found.
+* Improved error handling in the case a Docker container cannot be found.
+* Added a check for if a build-magic container is already running and raises an exception if so.
 
 ## Version 0.3.0